migrated from Apollo Boost
