+++ client/package.json
@@ -5,9 +5,15 @@
   "dependencies": {
     "@reach/router": "^1.2.1",
     "apollo-boost": "^0.4.3",
+    "apollo-cache-inmemory": "^1.6.2",
+    "apollo-client": "^2.6.3",
+    "apollo-link": "^1.2.12",
+    "apollo-link-error": "^1.1.11",
+    "apollo-link-http": "^1.5.15",
     "date-fns": "^1.30.1",
     "govuk-frontend": "^2.13.0",
     "graphql": "^14.4.1",
+    "graphql-tag": "^2.10.1",
     "query-string": "^6.8.1",
     "react": "^16.8.6",
     "react-apollo": "^2.5.8",
