+++ client/src/index.js
@@ -1,33 +1,32 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import ApolloClient from 'apollo-boost';
-import { ApolloProvider } from 'react-apollo';
 
-import * as serviceWorker from './serviceWorker';
+import { ApolloClient } from 'apollo-client';
+import { InMemoryCache } from 'apollo-cache-inmemory';
+import { HttpLink } from 'apollo-link-http';
+import { onError } from 'apollo-link-error';
+import { ApolloLink } from 'apollo-link';
+import { ApolloProvider } from 'react-apollo';
 
 import App from './App';
 
 const client = new ApolloClient({
-  uri: 'http://localhost:4050/graphql',
-  fetchOptions: {
-    credentials: 'include'
-  },
-  request: (operation) => {
-    const token = localStorage.getItem('rlbtoken');
-    operation.setContext({
-      headers: {
-        authorization: token
-      }
-    });
-  },
-  onError: ({ networkError }) => {
-    if (networkError) {
-      console.log('Network error', networkError);
-      if (networkError.status === 401) {
-        localStorage.removeItem('rlbtoken');
-      }
-    }
-  }
+  link: ApolloLink.from([
+    onError(({ graphQLErrors, networkError }) => {
+      if (graphQLErrors)
+        graphQLErrors.map(({ message, locations, path }) =>
+          console.log(
+            `[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`
+          )
+        );
+      if (networkError) console.log(`[Network error]: ${networkError}`);
+    }),
+    new HttpLink({
+      uri: 'http://localhost:4050/graphql',
+      credentials: 'same-origin'
+    })
+  ]),
+  cache: new InMemoryCache()
 });
 
 ReactDOM.render(
@@ -37,8 +36,3 @@ ReactDOM.render(
 
   document.getElementById('root')
 );
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: http://bit.ly/CRA-PWA
-serviceWorker.unregister();
